test: moving from Main to ToDoListFragment
