+++ app/src/main/res/values/strings.xml
@@ -16,4 +16,5 @@
     <string name="add_task">Add task</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="addtodo">AddTODO</string>
 </resources>
